requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-extensions==1.7.6
django-filter==1.0.1
django==1.10.5
djangorestframework==3.5.4
markdown==2.6.8
pygraphviz==1.3.1
six==1.10.0               # via django-extensions
Werkzeug==0.11.15
